API: libpe_status,libpe_rules,libcrmcommon: move enum rsc_recovery_type to libcrmcommon
This creates two new header files for libcrmcommon, resources.h (which will
have scheduler APIs related to resources) and scheduler.h (which as of this
commit does nothing but include resources.h, but will eventually include all
scheduler APIs so source files can include just scheduler.h).

enum rsc_recovery_type is moved from include/crm/pengine/common.h to
include/crm/common/resources.h, as the start of a project to merge all
functionality of libpe_status and libpe_rules into libcrmcommon.
